Removed unused method receiver; Removed Unused Code; Removed serving of OpenAPI
